+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,8 @@
+package command;
+import components.Storage;
+import components.Ui;
+import task.TaskList;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
         super.isExit = true;
