+++ src/main/java/command/LoadCommand.java
@@ -1,3 +1,9 @@
+package command;
+import components.Storage;
+import components.Ui;
+import exception.NyabotException;
+import task.TaskList;
+
 public class LoadCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws NyabotException {
