+++ src/main/java/components/Ui.java
@@ -1,3 +1,5 @@
+package components;
+
 import java.util.Scanner;
 
 public class Ui {
