+++ src/main/java/exception/NyabotDateException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class NyabotDateException extends NyabotException {
     public NyabotDateException(String str) {
         super(str);
