+++ src/main/java/exception/NyabotException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class NyabotException extends Exception {
     public NyabotException(String str) {
         super(str);
