+++ src/main/java/exception/NyabotIndexOutOfBoundsException.java
@@ -1,4 +1,4 @@
-import java.util.NoSuchElementException;
+package exception;
 
 public class NyabotIndexOutOfBoundsException extends NyabotException {
     public NyabotIndexOutOfBoundsException(String msg) {
