+++ src/main/java/exception/NyabotMissingArgumentException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class NyabotMissingArgumentException extends NyabotException {
     public NyabotMissingArgumentException(String missingArgument) {
         super(missingArgument);
