+++ src/main/java/exception/NyabotNoSuchCommandException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class NyabotNoSuchCommandException extends NyabotException {
     public NyabotNoSuchCommandException(String msg) {
         super(msg);
